Handle cases where native API is present but unavailable
The isPresent test only lets us know whether the native getFromLocation and getFromLocationName APIs are
available, but doesn't say anything about the APIs ability to return results. In cases where either of these APIs fail,
despite being present, an empty list is returned, which is then passed off as a success to the higher levels. From
the flutter client side, this then results in a null gRPC call, despite having otherwise succeeded at the lower levels
and consequently having failed to raise any exceptions.

The simplest solution, and the one used here, is to test the number of addresses returned by the native APIs,
and in cases where this is 0, simply throw a NotAvailableException.

A more elaborate solution would be to check the number of addresses returned, and fall back on an alternative
method of lookup, but this is left as future work.